backend/requirements.txt
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
av==14.4.0
babel==2.17.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-default-group==1.2.4
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
condense-json==0.1.3
confection==0.1.5
cryptography==45.0.4
csvw==3.5.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
dacite==1.9.2
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
distro==1.9.0
dlinfo==2.0.0
dnspython==2.7.0
docopt==0.6.2
einops==0.8.1
einx==0.3.0
espeakng-loader==0.2.4
fastapi==0.115.14
fastrtc==0.0.28
fastrtc-moonshine-onnx==20241016
ffmpy==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
google-crc32c==1.7.1
gradio==5.35.0
gradio_client==1.10.4
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llm==0.26
llm-mlx==0.4
llvmlite==0.44.0
loguru==0.7.3
macholib==1.16.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
misaki==0.9.4
mlx==0.26.1
mlx-audio==0.2.3
mlx-lm==0.25.3
mlx-vlm==0.1.27
modulegraph==0.19.6
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
networkx==3.5
num2words==0.5.14
numba==0.61.2
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.22.0
openai==1.93.0
opencv-python==4.10.0.84
orjson==3.10.18
packaging==25.0
pandas==2.3.0
phonemizer-fork==3.3.2
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
plyer==2.1.0
pooch==1.8.2
preshed==3.0.10
propcache==0.3.2
protobuf==6.31.1
puremagic==1.29
py2app==0.28.8
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.5
pylibsrtp==0.12.0
pyloudnorm==0.1.1
pyobjus==1.2.3
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-ulid==3.0.0
pytz==2025.2
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.26.0
ruff==0.12.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.13.1
segments==2.3.0
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.38
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
webrtcvad==2.0.10
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
